Separated logic of commands to modules under commands sub-dir
